requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via
    #   fsspec
    #   litellm
aiosignal==1.3.2
    # via aiohttp
alembic==1.15.2
    # via optuna
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   asyncer
    #   dspy
    #   google-genai
    #   httpx
    #   openai
asyncer==0.0.8
    # via dspy
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via dspy
blinker==1.9.0
    # via streamlit
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via
    #   dspy
    #   google-auth
    #   streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   litellm
    #   pip-tools
    #   streamlit
cloudpickle==3.1.1
    # via dspy
colorlog==6.9.0
    # via optuna
datasets==3.6.0
    # via dspy
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via dspy
distro==1.9.0
    # via openai
dspy==2.6.14
    # via -r requirements.in
filelock==3.18.0
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.3.0
    # via
    #   datasets
    #   huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-auth==2.40.1
    # via google-genai
google-genai==1.16.1
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   litellm
    #   openai
huggingface-hub==0.31.4
    # via
    #   datasets
    #   tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via
    #   altair
    #   litellm
    #   pydeck
jiter==0.10.0
    # via openai
joblib==1.5.0
    # via dspy
json-repair==0.45.1
    # via dspy
jsonschema==4.23.0
    # via
    #   altair
    #   litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.63.7
    # via dspy
magicattr==0.1.6
    # via dspy
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
narwhals==1.40.0
    # via altair
numpy==2.2.6
    # via
    #   datasets
    #   dspy
    #   optuna
    #   pandas
    #   pydeck
    #   streamlit
openai==1.61.0
    # via
    #   dspy
    #   litellm
optuna==4.3.0
    # via dspy
packaging==24.2
    # via
    #   altair
    #   build
    #   datasets
    #   huggingface-hub
    #   optuna
    #   pytest
    #   streamlit
pandas==2.2.3
    # via
    #   datasets
    #   dspy
    #   streamlit
pillow==11.2.1
    # via streamlit
pip-tools==7.4.1
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.0
    # via streamlit
pyarrow==20.0.0
    # via
    #   datasets
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via
    #   dspy
    #   google-genai
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.5
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   litellm
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
    #   optuna
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   dspy
    #   tiktoken
requests==2.32.3
    # via
    #   datasets
    #   dspy
    #   google-genai
    #   huggingface-hub
    #   streamlit
    #   tiktoken
rich==14.0.0
    # via dspy
rpds-py==0.25.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
ruff==0.11.10
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via
    #   alembic
    #   optuna
streamlit==1.45.1
    # via -r requirements.in
tenacity==9.1.2
    # via
    #   dspy
    #   streamlit
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.1
    # via litellm
toml==0.10.2
    # via streamlit
tornado==6.5
    # via streamlit
tqdm==4.67.1
    # via
    #   datasets
    #   dspy
    #   huggingface-hub
    #   openai
    #   optuna
typing-extensions==4.13.2
    # via
    #   alembic
    #   altair
    #   google-genai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
ujson==5.10.0
    # via dspy
urllib3==2.4.0
    # via requests
websockets==15.0.1
    # via google-genai
wheel==0.45.1
    # via pip-tools
xxhash==3.5.0
    # via datasets
yarl==1.20.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
